py3 compatibility: ensure that indeces are integers
1000/2 = 500 in py2
1000/2 = 500. in py3

+++ __init__.py
@@ -46,8 +46,8 @@ class ConfusionMatrixAnalyser(object):
 
     def gelman_rubin_test_on_samples(self, samples):
         no_samples = len(samples)
-        split_samples = np.stack([samples[:(no_samples / 2)],
-                                     samples[(no_samples / 2):]])
+        split_samples = np.stack([samples[:int(no_samples / 2)],
+                                  samples[int(no_samples / 2):]])
         passed_gelman_rubin = (pm.diagnostics.gelman_rubin(split_samples) < 1.01).all()
         return passed_gelman_rubin
